% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraping.R
\name{compareFit}
\alias{compareFit}
\title{compareFit}
\usage{
compareFit(data, fit, col = NULL, digits = 4, ...)
}
\arguments{
\item{data}{The data set, of class \code{dataFrap.class}.}

\item{fit}{A single fit object or a list of fit objects obtained from the \code{newFit} function.}

\item{col}{Colors to use for the plots. If NULL, colors are automatically assigned.}

\item{digits}{Number of digits to round the error values. Default is 4.}

\item{...}{Additional parameters passed to the underlying functions.}
}
\description{
Compare the fit of multiple models using histograms and Q-Q plots.
}
\examples{
# Compare the fit of multiple models using histograms and Q-Q plots
compareFit(data, fit = list(fit1, fit2, fit3))

}
\keyword{Compare}
\keyword{Fit}
\keyword{Q-Q}
\keyword{histograms}
\keyword{models}
\keyword{plot}
