% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraping.R
\name{compareMean}
\alias{compareMean}
\title{compareMean}
\usage{
compareMean(
  data1,
  data2,
  fit,
  type = "RCB",
  area = T,
  stand = T,
  simulated = F,
  Nsim = 50,
  seed = NA,
  saveable = !is.na(seed),
  conf.level = 0.95,
  alternative = "two.sided",
  return = F,
  p.value = T,
  npoints = 100,
  new.plot = T,
  plot.lines = T,
  plot.points = F,
  col = NULL,
  ...
)
}
\arguments{
\item{data1}{The first data set, of class \code{dataFrap.class}.}

\item{data2}{The second data set, of class \code{dataFrap.class}.}

\item{fit}{An optional fit object obtained from the \code{newFit} function.}

\item{type}{A vector or list of recovery curve types. Default is "RCB".}

\item{area}{A vector or list indicating whether to use area normalization. Default is TRUE.}

\item{stand}{A vector or list indicating whether to apply standardization. Default is TRUE.}

\item{simulated}{A vector or list indicating whether to perform simulations. Default is FALSE.}

\item{Nsim}{A vector or list specifying the number of simulations to perform. Default is 50.}

\item{seed}{A vector or list specifying the seed for the random number generator. Default is NA (no seed).}

\item{saveable}{A vector or list indicating whether to save the simulation results. Default is TRUE if seed is provided, otherwise FALSE.}

\item{conf.level}{Confidence level for hypothesis testing. Default is 0.95.}

\item{alternative}{The alternative hypothesis for hypothesis testing. Default is "two.sided".}

\item{return}{Boolean indicating whether to return the test results. Default is FALSE.}

\item{p.value}{Boolean indicating whether to plot the p-value. Default is TRUE.}

\item{npoints}{Number of points to use for generating the comparison plot. Default is 100.}

\item{new.plot}{Boolean indicating whether to create a new plot. Default is TRUE.}

\item{plot.lines}{Boolean indicating whether to plot the comparison as lines. Default is TRUE.}

\item{plot.points}{Boolean indicating whether to plot the comparison as points. Default is FALSE.}

\item{col}{Colors to use for the plot. If NULL, colors are automatically assigned.}

\item{...}{Additional plot parameters.}
}
\value{
If \code{return} is TRUE, returns the test results. Otherwise, generates the comparison plot.
}
\description{
Compare the mean recovery curves between two data sets.
}
\examples{
# Compare the mean recovery curves between two data sets
compareMean(data1, data2, fit = myFit)

# Compare the mean recovery curves and return the test results
result <- compareMean(data1, data2, fit = myFit, return = TRUE)
print(result)

}
\keyword{Compare}
\keyword{Mean}
\keyword{curves}
\keyword{data}
\keyword{hypothesis}
\keyword{recovery}
\keyword{sets}
\keyword{test}
