% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraping.R
\name{compareParam}
\alias{compareParam}
\title{compareParam}
\usage{
compareParam(
  data1,
  data2,
  fit,
  param,
  type = "RCB",
  area = T,
  stand = T,
  simulated = F,
  Nsim = 50,
  seed = NA,
  conf.level = 0.95,
  alternative = "two.sided",
  return = F,
  new.plot = T,
  plot.lines = T,
  plot.points = F,
  col = NULL,
  ...
)
}
\arguments{
\item{data1}{The first data set, of class \code{dataFrap.class}.}

\item{data2}{The second data set, of class \code{dataFrap.class}.}

\item{fit}{An optional fit object obtained from the \code{newFit} function.}

\item{param}{The name of the parameter to compare.}

\item{type}{A vector or list of recovery curve types. Default is "RCB".}

\item{area}{A vector or list indicating whether to use area normalization. Default is TRUE.}

\item{stand}{A vector or list indicating whether to apply standardization. Default is TRUE.}

\item{simulated}{A vector or list indicating whether to perform simulations. Default is FALSE.}

\item{Nsim}{A vector or list specifying the number of simulations to perform. Default is 50.}

\item{seed}{A vector or list specifying the seed for the random number generator. Default is NA (no seed).}

\item{conf.level}{Confidence level for hypothesis testing. Default is 0.95.}

\item{alternative}{The alternative hypothesis for hypothesis testing. Default is "two.sided".}

\item{return}{Boolean indicating whether to return the test results. Default is FALSE.}

\item{new.plot}{Boolean indicating whether to create a new plot. Default is TRUE.}

\item{plot.lines}{Boolean indicating whether to plot the comparison as lines. Default is TRUE.}

\item{plot.points}{Boolean indicating whether to plot the comparison as points. Default is FALSE.}

\item{col}{Colors to use for the plot. If NULL, colors are automatically assigned.}

\item{...}{Additional plot parameters.}
}
\value{
If \code{return} is TRUE, returns the test results. Otherwise, generates the comparison plot.
}
\description{
Compare a specific parameter between two data sets.
}
\examples{
# Compare the parameter between two data sets
compareParam(data1, data2, fit = myFit, param = "alpha")

# Compare the parameter and return the test results
result <- compareParam(data1, data2, fit = myFit, param = "alpha", return = TRUE)
print(result)

}
\keyword{Compare}
\keyword{Parameter}
\keyword{data}
\keyword{hypothesis}
\keyword{sets}
\keyword{test}
