% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraping.R
\name{.frapclass}
\alias{.frapclass}
\title{frapclass}
\format{
\code{R6Class} object.
}
\description{
Class representing functionality for FRAP (Fluorescence Recovery After Photobleaching).
}
\section{Usage}{

\preformatted{frap_obj <- frapclass$new(name, bleach)}
}

\examples{
# Create an instance of the "frapclass" class
frap_obj <- frapclass$new("FRAP1", 30)

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{Name of the class.}

\item{\code{bleach}}{Percentage of sample bleaching.}

\item{\code{fmin}}{Minimum fluorescence recovery value.}

\item{\code{length}}{Length of the class.}

\item{\code{seconds}}{Vector of time in seconds.}

\item{\code{photo}}{Vector of time frames.}

\item{\code{time0}}{Vector of recovery start times.}

\item{\code{getColor}}{Function to get the current color.}

\item{\code{setColor}}{Function to set the color.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-frapclass-new}{\code{.frapclass$new()}}
\item \href{#method-frapclass-addDataFrame}{\code{.frapclass$addDataFrame()}}
\item \href{#method-frapclass-addFile}{\code{.frapclass$addFile()}}
\item \href{#method-frapclass-addDir}{\code{.frapclass$addDir()}}
\item \href{#method-frapclass-time}{\code{.frapclass$time()}}
\item \href{#method-frapclass-area}{\code{.frapclass$area()}}
\item \href{#method-frapclass-recover}{\code{.frapclass$recover()}}
\item \href{#method-frapclass-table}{\code{.frapclass$table()}}
\item \href{#method-frapclass-clone}{\code{.frapclass$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-frapclass-new"></a>}}
\if{latex}{\out{\hypertarget{method-frapclass-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.frapclass$new(name, bleach)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-frapclass-addDataFrame"></a>}}
\if{latex}{\out{\hypertarget{method-frapclass-addDataFrame}{}}}
\subsection{Method \code{addDataFrame()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.frapclass$addDataFrame(
  seconds,
  photo,
  areaColumn,
  inputColumn,
  dataR,
  dataC,
  dataB = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-frapclass-addFile"></a>}}
\if{latex}{\out{\hypertarget{method-frapclass-addFile}{}}}
\subsection{Method \code{addFile()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.frapclass$addFile(
  seconds,
  photo,
  areaColumn,
  inputColumn,
  fileR,
  fileC,
  fileB = ""
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-frapclass-addDir"></a>}}
\if{latex}{\out{\hypertarget{method-frapclass-addDir}{}}}
\subsection{Method \code{addDir()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.frapclass$addDir(
  seconds,
  photo,
  areaColumn,
  inputColumn,
  dirR,
  dirC,
  dirB = ""
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-frapclass-time"></a>}}
\if{latex}{\out{\hypertarget{method-frapclass-time}{}}}
\subsection{Method \code{time()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.frapclass$time(index = NA, AB = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-frapclass-area"></a>}}
\if{latex}{\out{\hypertarget{method-frapclass-area}{}}}
\subsection{Method \code{area()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.frapclass$area(index = NA, type = "R")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-frapclass-recover"></a>}}
\if{latex}{\out{\hypertarget{method-frapclass-recover}{}}}
\subsection{Method \code{recover()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.frapclass$recover(
  index = NA,
  type = "RCB",
  area = TRUE,
  stand = TRUE,
  AB = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-frapclass-table"></a>}}
\if{latex}{\out{\hypertarget{method-frapclass-table}{}}}
\subsection{Method \code{table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.frapclass$table(
  index = NA,
  type = "RCB",
  area = TRUE,
  stand = TRUE,
  AB = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-frapclass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-frapclass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{.frapclass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
