% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{downloadData}
\alias{downloadData}
\title{Download test database}
\usage{
downloadData()
}
\description{
This function is used to download a test database for demonstration or testing purposes.
}
\details{
The \code{downloadData} function downloads a test database and does not perform any further operations on it.
}
\examples{
\dontrun{
  # Download the test database
  downloadData()
}

}
