% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraping.R
\name{newDataFrap}
\alias{newDataFrap}
\title{newDataFrap}
\usage{
newDataFrap(name, bleach)
}
\arguments{
\item{name}{Name of the FRAP object.}

\item{bleach}{Percentage of sample bleaching.}
}
\value{
An instance of the "frapclass" class.
}
\description{
Create a new instance of the "frapclass" class.
}
\examples{
# Create a new FRAP object
frap_obj <- newDataFrap("FRAP1", 30)

}
\keyword{FRAP}
\keyword{class}
\keyword{object}
