% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraping.R
\name{newFit}
\alias{newFit}
\title{newFit}
\usage{
newFit(name, fun, param, interval)
}
\arguments{
\item{name}{Name of the fit.}

\item{fun}{The fitting function.}

\item{param}{The parameter(s) of the fitting function.}

\item{interval}{The interval(s) for each parameter.}
}
\value{
An instance of the "Fit" class.
}
\description{
Create a new instance of the "Fit" class for a given function.
}
\examples{
# Create a new Fit object for the Exponential distribution
fit_exp <- newFit("Exponential", pexp, "rate", list(c(0, 1)))

# Create a new Fit object for the Pareto distribution
ppareto <- function(x, shape, scale) {
  if (any(x < scale)) return(rep(0, length(x)))
  shape * scale^shape / x^(shape + 1)
}
fit_pareto <- newFit("Pareto", ppareto, c("shape", "scale"), list(c(0, 5), c(0, 100)))

}
\keyword{Fit}
\keyword{class}
\keyword{object}
