% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraping.R
\name{plotFit}
\alias{plotFit}
\title{plotFit}
\usage{
plotFit(
  ...,
  fit,
  index = NA,
  type = "RCB",
  area = T,
  stand = T,
  simulated = F,
  Nsim = 50,
  seed = NA,
  npoints = 100,
  displacement = F,
  new.plot = T,
  plot.lines = F,
  plot.points = F,
  plot.shadow = F,
  plot.mean = F,
  col = NULL
)
}
\arguments{
\item{...}{One or more data sets to plot, each of class \code{dataFrap.class}.}

\item{fit}{An optional fit object obtained from the \code{newFit} function.}

\item{index}{Indices of the data sets to include in the plot. Default is NA (all data sets).}

\item{type}{A vector or list of recovery curve types. Default is "RCB".}

\item{area}{A vector or list indicating whether to use area normalization. Default is TRUE.}

\item{stand}{A vector or list indicating whether to apply standardization. Default is TRUE.}

\item{simulated}{A vector or list indicating whether to plot simulated recovery curves. Default is FALSE.}

\item{Nsim}{A vector or list specifying the number of simulations to perform. Default is 50.}

\item{seed}{A vector or list specifying the seed for the random number generator. Default is NA (no seed).}

\item{npoints}{Number of points to use for generating the fitted/simulated curves. Default is 100.}

\item{displacement}{Boolean indicating whether to displace the curves by their respective time0 values. Default is FALSE.}

\item{new.plot}{Boolean indicating whether to create a new plot. Default is TRUE.}

\item{plot.lines}{Boolean indicating whether to plot the fitted/simulated curves as lines. Default is FALSE.}

\item{plot.points}{Boolean indicating whether to plot the fitted/simulated curves as points. Default is FALSE.}

\item{plot.shadow}{Boolean indicating whether to plot the shadow area around the fitted/simulated curves. Default is FALSE.}

\item{plot.mean}{Boolean indicating whether to plot the mean fitted/simulated curve. Default is FALSE.}

\item{col}{Colors to use for the plot. If NULL, colors are automatically assigned.}
}
\description{
Plot the fitted or simulated recovery curves for multiple data sets.
}
\examples{
# Generate a plot of fitted recovery curves for all data sets
plotFit(data1, data2, data3, fit = myFit)

# Generate a plot of fitted recovery curves for specific data sets
plotFit(data1, data2, data3, fit = myFit, index = c(1, 3, 5))

# Generate a plot of simulated recovery curves
plotFit(data1, data2, data3, fit = myFit, simulated = TRUE, Nsim = 100)

}
\keyword{Fitted}
\keyword{Plot}
\keyword{curves}
\keyword{multiple}
\keyword{recovery}
\keyword{simulated}
