% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraping.R
\name{plotRecover}
\alias{plotRecover}
\title{plotRecover}
\usage{
plotRecover(
  ...,
  index = NA,
  type = "RCB",
  area = T,
  stand = T,
  AB = F,
  new.plot = T,
  plot.lines = F,
  plot.points = F,
  plot.shadow = F,
  plot.mean = F,
  col = NULL,
  getGroup = F
)
}
\arguments{
\item{...}{One or more data sets to plot, each of class \code{dataFrap.class}.}

\item{index}{Indices of the data sets to include in the plot. Default is NA (all data sets).}

\item{type}{A vector or list of recovery curve types. Default is "RCB".}

\item{area}{A vector or list indicating whether to use area normalization. Default is TRUE.}

\item{stand}{A vector or list indicating whether to apply standardization. Default is TRUE.}

\item{AB}{A vector or list indicating whether to include pre-bleach data. Default is FALSE.}

\item{new.plot}{Boolean indicating whether to create a new plot. Default is TRUE.}

\item{plot.lines}{Boolean indicating whether to plot recovery curves as lines. Default is FALSE.}

\item{plot.points}{Boolean indicating whether to plot recovery curves as points. Default is FALSE.}

\item{plot.shadow}{Boolean indicating whether to plot recovery curve shadows. Default is FALSE.}

\item{plot.mean}{Boolean indicating whether to plot the mean recovery curve. Default is FALSE.}

\item{col}{Colors to use for the plot. If NULL, colors are automatically assigned.}

\item{getGroup}{Boolean indicating whether to return the group function. Default is FALSE.}
}
\value{
If \code{getGroup} is TRUE, returns a function for grouping data. Otherwise, generates the plot.
}
\description{
Plot the recovery curves for multiple data sets.
}
\examples{
# Generate a plot of recovery curves for all data sets
plotRecover(data1, data2, data3)

# Generate a plot of recovery curves for specific data sets
plotRecover(data1, data2, data3, index = c(1, 3, 5))

# Return the group function for custom grouping
group_func <- plotRecover(getGroup = TRUE)
custom_group <- group_func(data1, data2)

}
\keyword{Plot}
\keyword{Recovery}
\keyword{curves}
\keyword{multiple}
