% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraping.R
\name{simFit}
\alias{simFit}
\title{simFit}
\usage{
simFit(fit, Nsim = 50, seed = NA, saveable = !is.na(seed))
}
\arguments{
\item{fit}{A Fit object obtained from the \code{newFit} function.}

\item{Nsim}{Number of simulations to perform.}

\item{seed}{Seed for reproducibility. Default is NA.}

\item{saveable}{Boolean indicating whether the simulation results should be saved. Default is !is.na(seed).}
}
\value{
A simulated Fit object.
}
\description{
Simulate fits based on a given Fit object.
}
\examples{
# Simulate fits for the Exponential distribution Fit object
sim_exp <- simFit(fit_exp, Nsim = 100, seed = 123)

# Simulate fits for the Pareto distribution Fit object
sim_pareto <- simFit(fit_pareto, Nsim = 50, seed = 456)

}
\keyword{Fit}
\keyword{Simulation}
\keyword{class}
\keyword{object}
