% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fraping.R
\name{tableFit}
\alias{tableFit}
\title{tableFit}
\usage{
tableFit(fit, index = NA, timelim = NULL, npoints = 100, displacement = F)
}
\arguments{
\item{fit}{A Fit object obtained from the \code{newFit} function.}

\item{index}{Indices of the fits to include in the table. Default is NA (all fits).}

\item{timelim}{Time limits for the table. Default is NULL (uses full time range).}

\item{npoints}{Number of points in the table. Default is 100.}

\item{displacement}{Boolean indicating whether to include time displacement. Default is FALSE.}
}
\value{
A table of fitted values.
}
\description{
Generate a table of fitted values based on a given Fit object.
}
\examples{
# Generate a table of fitted values for all fits in the Fit object
table_all <- tableFit(fit)

# Generate a table of fitted values for specific fits in the Fit object
table_subset <- tableFit(fit, index = c(1, 3, 5))

}
\keyword{Fit}
\keyword{Table}
\keyword{class}
\keyword{object}
