% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{test.fraping}
\alias{test.fraping}
\title{Example script for FRAP analysis}
\usage{
test.fraping()
}
\description{
This function provides an example script for FRAP (Fluorescence Recovery After Photobleaching) analysis using the 'fraping' package.
}
\details{
The \code{test.fraping} function demonstrates how to perform FRAP analysis on a set of data using the 'fraping' package. It includes the necessary steps to load the required libraries, set plot parameters, create data objects, add directories, and generate plots.
}
\examples{
\dontrun{
  # Run the example script
  test.fraping()
}

}
